Sp type bug.

+++ web-console/src/main/java/liquid/persistence/repository/SpRepository.java
@@ -10,5 +10,5 @@ import org.springframework.data.repository.CrudRepository;
  * Time: 5:03 PM
  */
 public interface SpRepository extends CrudRepository<ServiceProvider, Long> {
-    Iterable<ServiceProvider> findByType(int type);
+    Iterable<ServiceProvider> findByType(long type);
 }
